Explicitly call-out trailing slash in URL
Ran into an issue with a missing `/` at end of SAML URLs a bit ago. Search MacAdmins on this issue found some convos from the past few years so figured it might be beneficial to explicitly call-out that the trailing `/` at the end of these SAML URLs is important otherwise you end up debugging a `RuntimeError`.